backend/requirements.txt
fastapi
uvicorn
transformers
torch
datasets
pandas
faiss-cpu
